# Remove active_tasks counter from EpochManager