The counter can be a big bottleneck, as is visible from the following benchmarks performed on a Hash Table data structure I and @LouisJenkinsCS are working on:


**With `active_tasks` counter**
```
Intset Benchmark:
1 tasks, 347.816 ns/op
2 tasks, 306.365 ns/op
4 tasks, 251.258 ns/op
8 tasks, 262.818 ns/op
16 tasks, 204.799 ns/op
32 tasks, 210.513 ns/op
44 tasks, 131.005 ns/op
```


**Without `active_tasks` counter**
```
Intset Benchmark:
1 tasks, 333.38 ns/op
2 tasks, 212.159 ns/op
4 tasks, 114.36 ns/op
8 tasks, 59.4826 ns/op
16 tasks, 31.6458 ns/op
32 tasks, 19.9671 ns/op
44 tasks, 14.7049 ns/op
```

My initial intention to keep this counter was to speed up reclamation on nodes which did not have any pinned tasks, saving the locale the effort to scan through the list of tokens to detect if any task is active. However, as it is clear from the benchmarks, this can create serious bottlenecks, and the GC is much more efficient without the counter.