@ronawho 2019-10-05
Just curious -- is there less of a performance hit if you use a processor atomic? e.g. by switching to `var active_tasks : chpl__processorAtomicType(uint);`