@dgarvit 2019-10-06
I did run the previous benchmarks with `CHPL_NETWORK_ATOMICS=none`, so the benchmark with `var active_tasks : chpl__processorAtomicType(uint);` seems no different:
```
Intset Benchmark:
1 tasks, 346.828 ns/op
2 tasks, 252.094 ns/op
4 tasks, 260.756 ns/op
8 tasks, 239.798 ns/op
16 tasks, 212.334 ns/op
32 tasks, 159.21 ns/op
44 tasks, 142.832 ns/op
```